# swagger/admin/swagger.yaml
definitions:
  dto.ContactInfo:
    properties:
      bin:
        description: БИН организации
        example: "5189011425"
        type: string
      email:
        description: Имейл
        example: oleg@tinkoff.ru
        type: string
      first_name:
        description: Имя
        example: Oleg
        type: string
      last_name:
        description: Фамилия
        example: Tinkoff
        type: string
      location:
        description: Адрес организации
        example: Bldg. 26, 38A, 2 Khutorskaya str., Moscow, Russia.
        type: string
      phone:
        description: Телефон
        example: "87777777777"
        type: string
      web_address:
        description: Веб адрес организации
        example: https://www.tinkoff.ru
        type: string
    type: object
  dto.IDResponse:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
  dto.MarketRequest:
    properties:
      code:
        description: Код картинки
        example: tinkoff
        type: string
      company_name:
        description: Название компании
        type: string
      contact:
        $ref: '#/definitions/dto.ContactInfo'
        description: Контакт
      enabled:
        description: Включение
        example: true
        type: boolean
      logo_url:
        description: Веб адрес картинки
        example: https://plusworld.ru/wp-content/uploads/2021/09/759950f2eafdab128b887f1296316ea6f1af5152.jpg
        type: string
      password:
        description: Пароль
        example: Password123
        type: string
      update_order_url:
        description: Обновление заказов
        example: https://google.com
        type: string
      username:
        description: Имя пользователя
        example: Olejka
        type: string
    type: object
  dto.Markets:
    properties:
      markets:
        items:
          $ref: '#/definitions/models.Market'
        type: array
      total:
        type: integer
    type: object
  dto.OfferRequest:
    properties:
      max_order_sum:
        description: Максимальная сумма заказа
        type: integer
      min_order_sum:
        description: Минимальная сумма заказа
        type: integer
      name:
        description: Наименование
        example: МФО Аирба
        type: string
      partner_code:
        description: Код партнера
        example: airba_pay
        type: string
      payment_type_group_code:
        description: Вид оплаты товара
        example: online_broker
        type: string
    type: object
  dto.OfferSpecs:
    properties:
      offers:
        items:
          $ref: '#/definitions/models.Offer'
        type: array
      total:
        type: integer
    type: object
...
  httperrors.Details:
    properties:
      code:
        description: application-определенный код ошибки
        type: integer
      message:
        description: application-level сообщение, для дебага
        type: string
      status:
        description: сообщение пользовательского уровня
        type: string
    type: object
  httperrors.Response:
    properties:
      error:
        $ref: '#/definitions/httperrors.Details'
      validation:
        additionalProperties:
          type: string
        description: ошибки валидации
        type: object
    type: object
...
  models.ContactInfo:
    properties:
      bin:
        description: БИН организации
        example: "5189011425"
        type: string
      email:
        description: Имейл
        example: oleg@tinkoff.ru
        type: string
      first_name:
        description: Имя
        example: Oleg
        type: string
      last_name:
        description: Фамилия
        example: Tinkoff
        type: string
      location:
        description: Адрес организации
        example: Bldg. 26, 38A, 2 Khutorskaya str., Moscow, Russia.
        type: string
      phone:
        description: Телефон
        example: "87777777777"
        type: string
      web_address:
        description: Веб адрес организации
        example: https://www.tinkoff.ru
        type: string
    type: object
...
  models.Market:
    properties:
      code:
        type: string
      contact:
        $ref: '#/definitions/models.ContactInfo'
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      logo_url:
        type: string
      password:
        type: string
      title:
        type: string
      update_order_url:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.Offer:
    properties:
      created_at:
        type: string
      id:
        type: string
      max_order_sum:
        type: integer
      min_order_sum:
        type: integer
      name:
        type: string
      partner_code:
        type: string
      payment_type_group_code:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Partner:
    properties:
      code:
        type: string
      commission:
        type: integer
      company_name:
        type: string
      contact:
        $ref: '#/definitions/models.ContactInfo'
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      logo_url:
        type: string
      password:
        type: string
      updated_at:
        type: string
      url:
        $ref: '#/definitions/models.URL'
      username:
        type: string
    type: object
...
  models.Response:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
  models.URL:
    properties:
      create:
        type: string
      update:
        type: string
    type: object
info:
  contact: {}
paths:
  /admins/markets:
    get:
      description: Получение маркетов
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: pagination limit
        in: query
        name: limit
        type: integer
      - description: pagination skip
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Markets'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Получение маркетов
      tags:
      - Market
    post:
      consumes:
      - application/json
      description: Создание нового маркета
      parameters:
      - description: body
        in: body
        name: market
        required: true
        schema:
          $ref: '#/definitions/dto.MarketRequest'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Создание нового маркета
      tags:
      - Market
  /admins/markets/{id}:
    get:
      consumes:
      - application/json
      description: Получение маркета
      parameters:
      - description: id of the partner
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Market'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Получение маркета
      tags:
      - Market
    put:
      description: Обновление определенного маркета
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: body
        in: body
        name: market
        required: true
        schema:
          $ref: '#/definitions/dto.MarketRequest'
      - description: id of the partner
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Обновление определенного маркета
      tags:
      - Market
  /admins/offers:
    get:
      consumes:
      - application/json
      description: Получение офферов
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: pagination limit
        in: query
        name: limit
        type: integer
      - description: pagination skip
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OfferSpecs'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Получение офферов
      tags:
      - Offers
    post:
      consumes:
      - application/json
      description: Создание нового оффера
      parameters:
      - description: body
        in: body
        name: partner
        required: true
        schema:
          $ref: '#/definitions/dto.OfferRequest'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Создание нового оффера
      tags:
      - Offers
  /admins/offers/{id}:
    get:
      consumes:
      - application/json
      description: Получение оффера
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: id of the market
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Partner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Получение оффера
      tags:
      - Offers
    put:
      consumes:
      - application/json
      description: Обновление оффера
      parameters:
      - description: body
        in: body
        name: partner
        required: true
        schema:
          $ref: '#/definitions/dto.OfferRequest'
      - description: id of the market
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Offer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Обновление оффера
      tags:
      - Offers
...
swagger: "2.0"